% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d3scatter.R
\name{d3scatter}
\alias{d3scatter}
\title{Create a scatter plot}
\usage{
d3scatter(data, x, y, color = NULL, x_label = NULL, y_label = NULL,
  x_lim = NULL, y_lim = NULL, width = NULL, height = NULL)
}
\arguments{
\item{data}{A data frame or \link[crosstalk]{SharedData} object}

\item{x}{One-sided formula indicating the column or expression for x values}

\item{y}{One-sided formula indicating the column or expression for y values}

\item{color}{One-sided formula indicating the column or expression for color
values}

\item{x_label, y_label}{Labels for axes}

\item{x_lim, y_lim}{Two-element numeric vectors indicating the limits for axes}

\item{width, height}{Override default size (see
\link[htmltools]{validateCssUnit})}
}
\description{
Create a scatter plot
}
\examples{
d3scatter(iris, ~Sepal.Width, ~Sepal.Length, color = ~Species)

}

